% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_collapse_right}
\alias{array_collapse_right}
\title{array_collapse_right}
\usage{
array_collapse_right(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes n x m x k.}

\item{B}{A matrix with shapes 1 x n.}
}
\description{
Calculates the matrix product between an array and a vector.
}
\details{
For an array A with shapes m x n x k and a vector B with shape m, this operations returns a matrix C, with shapes n x k, so that C[,i] = B %*% A[,,i].
}
\keyword{internal}
