% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_mult_left}
\alias{array_mult_left}
\title{array_mult_left}
\usage{
array_mult_left(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes n x m x k.}

\item{B}{A matrix with shapes m x l.}
}
\description{
Calculates the matrix product between an array and a matrix.
}
\details{
For an array A with shapes n x m x k and a matrix B with shape m x l, this operations returns an array C, with shapes n x l x k, so that C[,,i] = A[,,i] %*% B.
}
\keyword{internal}
