% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_mult_right}
\alias{array_mult_right}
\title{array_mult_right}
\usage{
array_mult_right(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes n x m x k.}

\item{B}{A matrix with shapes l x n.}
}
\description{
Calculates the matrix product between an array and a matrix.
}
\details{
For an array A with shapes m x n x k and a matrix B with shape l x m, this operations returns an array C, with shapes l x n x k, so that C[,,i] = B %*% A[,,i].
}
\keyword{internal}
