% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_transp}
\alias{array_transp}
\title{array_transp}
\usage{
array_transp(A)
}
\arguments{
\item{A}{A 3-D array.}
}
\description{
Calculates the element-wise transposition of an array.
}
\details{
For an array A with shapes n x m x k, this operations returns an array C, with shapes m x n x k, so that C[,,i] = t(A[,,i]).
}
\keyword{internal}
