% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{base_ribbon}
\alias{base_ribbon}
\title{base_ribbon}
\usage{
base_ribbon(x, ymin, ymax, ...)
}
\arguments{
\item{x}{numeric: A sequence of values for the x-axis.}

\item{ymin}{numeric: A sequence of values for lower bound of the ribbon.}

\item{ymax}{numeric: A sequence of values for upper bound of the ribbon.}

\item{...}{Extra arguments for the polygon function.}
}
\description{
Makes a ribbon plot using R base functions.
}
\keyword{internal}
