% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{bdiag}
\alias{bdiag}
\title{bdiag}
\usage{
bdiag(...)
}
\arguments{
\item{...}{A list of matrices to be used.}
}
\value{
A block diagonal matrix whose diagonal elements are equal to the matrices passed as arguments.
}
\description{
Creates a block diagonal matrix with the matrix passed as argument.
}
\keyword{internal}
