% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{calcula_max}
\alias{calcula_max}
\title{calcula_max}
\usage{
calcula_max(pre_max)
}
\arguments{
\item{pre_max}{Numeric: A vector/matrix from which to calculate the axis limits and gradation.}
}
\value{
A list contaning the gradation for the axis, the number of ticks in the axis and the maximum value.
}
\description{
Auxiliary function to calculate the axis limits and gradation for plots.
}
\keyword{internal}
