% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{CAR_prior}
\alias{CAR_prior}
\title{CAR prior}
\usage{
CAR_prior(block, adj.matrix, scale, rho, var.index = 1:block$n)
}
\arguments{
\item{block}{dlm_block: The structural block.}

\item{adj.matrix}{Matrix: The adjacency matrix.}

\item{rho}{Numeric: The rho parameter for the CAR model (see references).}

\item{tau}{Numeric: The tau parameter for the CAR model (see references).}
}
\value{
A dlm_block object with the desired prior.
}
\description{
Set the prior of a structural block as a Conditional Autoregressive (CAR) prior.
}
\details{
For a revision of the CAR prior, see \insertCite{AlexCar;textual}{kDGLM}.

For the details about the implementation see \insertCite{ArtigoPacote;textual}{kDGLM}.
}
\references{
\insertAllCited{}
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{*.dlm_block}()},
\code{\link{+.dlm_block}()},
\code{\link{AR_block}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{print.dlm_block}()},
\code{\link{regression_block}()},
\code{\link{summary.dlm_block}()},
\code{\link{zero_sum_prior}()}
}
\concept{{auxiliary functions for structural blocks}}
