% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{check.block.status}
\alias{check.block.status}
\title{check.block.status}
\usage{
check.block.status(block)
}
\arguments{
\item{block}{A dlm_block object.}
}
\value{
A character ("defined" or "undefined") indicating if all parameters in the block are defined.
}
\description{
Checks if a block is defined.
}
\keyword{internal}
