% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{check.expr}
\alias{check.expr}
\title{check.expr}
\usage{
check.expr(expr)
}
\arguments{
\item{expr}{A string or expression.}
}
\description{
Checks if a string is expression and, if it is, evaluates it.
}
\keyword{internal}
