% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chickenPox}
\alias{chickenPox}
\title{Hospital admissions by chicken pox in Brazil}
\format{
A data frame with 120 rows and 6 columns:
\describe{
  \item{date}{The date of the observations.}
  \item{< 5 year, 5 to 9 years, 10 to 14 years, 15 to 49 years, 50 years or more}{The number of admissions for each age group.}
}
}
\source{
\url{https://datasus.saude.gov.br/informacoes-de-saude-tabnet/}
}
\usage{
chickenPox
}
\description{
Monthly hospital admissions by chicken pox in Brazil from January 2010 to December 2019.
}
\keyword{datasets}
