% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coefficients.fitted_dlm}
\alias{coefficients.fitted_dlm}
\title{coefficients.fitted_dlm}
\usage{
\method{coefficients}{fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to coef.}
}
\description{
This method is wrapper for the coef method.
}
\seealso{
\code{\link{coef.fitted_dlm}}
}
\keyword{internal}
