% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{convert_FGamma_Normal}
\alias{convert_FGamma_Normal}
\title{convert_FGamma_Normal}
\usage{
convert_FGamma_Normal(ft, Qt, parms)
}
\arguments{
\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this function.}
}
\value{
The parameters of the conjugated distribution of the linear predictor.
}
\description{
Calculate the parameters of the conjugated prior that best approximates the given log-Normal distribution.
The approximation is the best in the sense that it minimizes the KL divergence from the log-Normal to the conjugated prior.
}
\seealso{
Other {auxiliary functions for a Gamma outcome with unknowned shape}: 
\code{\link{Fgamma_pred_alt}()},
\code{\link{Fgamma_pred}()},
\code{\link{convert_Normal_FGamma}()},
\code{\link{update_FGamma_alt}()},
\code{\link{update_FGamma}()}
}
\concept{{auxiliary functions for a Gamma outcome with unknowned shape}}
\keyword{internal}
