% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multnom.R
\name{convert_Normal_Multinom}
\alias{convert_Normal_Multinom}
\title{convert_Normal_Multinom}
\usage{
convert_Normal_Multinom(conj.param, parms = list())
}
\arguments{
\item{conj.param}{list: A vector containing the concentration parameters of the Dirichlet.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this kernel.}
}
\value{
The parameters of the Normal distribution of the linear predictor.
}
\description{
Calculate the parameters of the log-Normal that best approximates the given Dirichlet distribution.
The approximation is the best in the sense that it minimizes the KL divergence from the Dirichlet to the log-Normal
}
\seealso{
Other auxiliary functions for a Multinomial outcome: 
\code{\link{convert_Multinom_Normal}()},
\code{\link{multnom_pred}()},
\code{\link{update_Multinom}()}
}
\concept{auxiliary functions for a Multinomial outcome}
\keyword{internal}
