% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_poisson.R
\name{convert_Normal_Poisson}
\alias{convert_Normal_Poisson}
\title{convert_Normal_Poisson}
\usage{
convert_Normal_Poisson(conj.param, parms)
}
\arguments{
\item{conj.param}{list: A vector containing the parameters of the Gamma (alpha,beta).}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this kernel.}
}
\value{
The parameters of the Normal distribution of the linear predictor.
}
\description{
Calculate the parameters of the log-Normal that best approximates the given Gamma distribution.
The approximation is the best in the sense that it minimizes the KL divergence from the Gamma to the log-Normal
}
\seealso{
Other auxiliary functions for a Poisson outcome: 
\code{\link{convert_Poisson_Normal}()},
\code{\link{poisson_pred}()},
\code{\link{update_Poisson}()}
}
\concept{auxiliary functions for a Poisson outcome}
\keyword{internal}
