% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cornWheat}
\alias{cornWheat}
\title{Corn and wheat prices from 1986 to 2014}
\format{
A data frame with 7,253 rows and 5 columns:
\describe{
  \item{date}{The date of the observation.}
  \item{corn.price, wheat.price}{ The price (in U.S. Dollars) per bushel of corn and wheat, respectively.}
  \item{corn.log.return, wheat.log.return }{The log returns for corn and wheat, respectively.}
}
}
\source{
\url{https://www.macrotrends.net/charts/commodities}
}
\usage{
cornWheat
}
\description{
The to prices (in U.S. Dollars) per bushel and the log returns of corn and wheat from 1986-01-03 to 2014-10-10.
Each observation corresponds to the price on that day, but not all days are present in this dataset.
}
\keyword{datasets}
