% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{create.multi.block}
\alias{create.multi.block}
\title{create.multi.block}
\usage{
create.multi.block(X, block.func)
}
\description{
Creates replicates of a block for each variable defined by X (either by formula or by a value)
}
\keyword{internal}
