% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{create_G}
\alias{create_G}
\title{create_G}
\usage{
create_G(S0, S1)
}
\arguments{
\item{S0}{A covariance matrix}

\item{S1}{Another covariance matrix}
}
\description{
Creates a matrix G such that G %*% S0 %*% t(G)= S1.
}
\keyword{internal}
