% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dglm.r
\name{dglm}
\alias{dglm}
\title{Create a Supervised Deep Learning Model}
\usage{
dglm(
  data,
  formula,
  hidden_layers = integer(),
  hidden_layers_activation = rep("linear", length(hidden_layers)),
  hidden_layer_names = paste("hidden_layer", seq_along(hidden_layers), sep = "_"),
  use_bias = rep(TRUE, length(hidden_layers)),
  loss = NULL,
  optimizer = optimizer_adadelta(),
  metrics = NULL,
  output_activation = NULL,
  output_activation_bias = TRUE,
  batch_size = nrow(data),
  epochs = 1000,
  verbose = FALSE,
  validation_split = 0.2,
  count_model = FALSE,
  name = NULL
)
}
\description{
Create a Supervised Deep Learning Model
}
