% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dlm_sampling}
\alias{dlm_sampling}
\title{dlm_sampling}
\usage{
dlm_sampling(model, sample.size, filt.distr = FALSE)
}
\arguments{
\item{model}{fitted_dlm: A fitted model from which to sample.}

\item{sample.size}{integer: The number of samples to draw.}

\item{filt.distr}{bool: A flag indicating if the samples should be dawned from the filtered (if TRUE) or smoothed distribution. The default is FALSE.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Array: An array containing the samples of latent states. Dimensions are n x t x sample.size, where n is the number of latent variables in the model and t is the number of observed values.
   \item ft Array: An array containing the samples of linear predictors. Dimensions are k x t x sample.size, where k is the number of linear predictors in the model and t is the number of observed values.
   \item param List: A named list containing, for each model outcome, an array with the samples of the parameters of the observational model. Each array will have dimensions l x t x sample.size, where l is the number of parameters in the observational model and t is the number of observed values.
}
}
\description{
This is function draws samples from the latent states using the backward sampling algorithm. See \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 15, for details.
}
\examples{

structure <- polynomial_block(mu = 1, D = 0.95) +
  polynomial_block(V = 1, D = 0.95)

outcome <- Normal(mu = "mu", V = "V", data = cornWheat$corn.log.return[1:500])
fitted.data <- fit_model(structure, outcome)

sample <- dlm_sampling(fitted.data, 5000)

}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{coef.fitted_dlm}()},
\code{\link{coefficients.fitted_dlm}()},
\code{\link{forecast.fitted_dlm}()},
\code{\link{update.fitted_dlm}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
