% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr.r
\name{dlr}
\alias{dlr}
\title{Create a Supervised Deep Learning Model}
\usage{
dlr(
  data,
  form,
  hidden_layers = integer(),
  hidden_layers_activation = rep("linear", length(hidden_layers)),
  categorical_loss = "mse",
  continuous_loss = "mse",
  optimizer = optimizer_adadelta(),
  metrics = c("accuracy", "mean_squared_error"),
  batch_size = nrow(data),
  epochs = 1000,
  verbose = FALSE,
  validation_split = 0.2,
  name = NULL
)
}
\description{
Create a Supervised Deep Learning Model
}
