% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{dmvnorm}
\alias{dmvnorm}
\title{dmvnorm}
\usage{
dmvnorm(x, mu, Sigma)
}
\arguments{
\item{x}{Vector: The value from to which calculate the density.}

\item{mu}{Vector: The mean vector}

\item{Sigma}{Matrix: The Covariance matrix.}
}
\description{
Calculates the log density of a multivariate normal distribution with mean mu and covariance matrix Sigma.
}
\keyword{internal}
