% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{effects.fitted_dlm}
\alias{effects.fitted_dlm}
\title{effects.fitted_dlm}
\usage{
\method{effects}{fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to plot_lat_var.}
}
\description{
effects method for class fitted_dlm
}
\keyword{internal}
