% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eval_past}
\alias{eval_past}
\title{eval_past}
\usage{
eval_past(model, smooth = FALSE, h = 0, pred_cred = 0.95)
}
\arguments{
\item{model}{<undefined class> or list: The fitted model to be use for evaluation.}

\item{smooth}{bool: The flag indicating if smoothed values should be used. If TRUE, h will not be used.}

\item{h}{positive integer: The relative offset for forecast. Values for time t will be calculated based on the filtered values of time t-h Will be ignored if smooth is TRUE.}

\item{pred_cred}{Numeric: The credibility level for the I.C. intervals.}
}
\value{
A list containing:
\itemize{
   \item pred Matrix: A matrix with the predictive mean at each time. Dimensions are k x t, where k is the number of outcomes.
   \item var.pred Array: A 3D-array with the predictive covariance matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item icl.pred Matrix: A matrix with the lower bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item icu.pred Matrix: A matrix with the upper bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
}
}
\description{
Evaluates the predictive values for the observed values used to fit the model. Predictions can be made with smoothed values or with filtered values with a time offset.
}
\examples{
T <- 200
w <- ((T + 20) / 40) * 2 * pi
y1 <- matrix(rpois((T + 20), 20 * (sin(w * 1:(T + 20) / (T + 20)) + 2)), (T + 20), 1)
y2 <- matrix(rpois((T + 20), 1:(T + 20) / (T + 20) + 1), (T + 20), 1)
y3 <- matrix(rpois((T + 20), 6), (T + 20), 1)
y <- cbind(y1, y2, y3)
y_pred <- y[T:(T + 20), ]

y <- y[1:T, ]

level <- polynomial_block(p1 = 1) + polynomial_block(p2 = 1)
season_2 <- harmonic_block(p2 = 1, period = 20)

outcome <- Multinom(p = c("p1", "p2"), outcome = y)

fitted_data <- fit_model(level, season_2, outcomes = outcome)

past <- eval_past(fitted_data, smooth = TRUE)

}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{dlm_sampling}()},
\code{\link{fit_model}()},
\code{\link{forecast}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
