% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{evaluate_max}
\alias{evaluate_max}
\title{evaluate_max}
\usage{
evaluate_max(pre.max)
}
\arguments{
\item{pre.max}{numeric: A vector/matrix from which to calculate the axis limits and gradation.}
}
\value{
A list containing the gradation for the axis, the number of ticks in the axis and the maximum value.
}
\description{
Auxiliary function to calculate the axis limits and gradation for plots.
}
\keyword{internal}
