% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{f_joint_root}
\alias{f_joint_root}
\title{f_joint_root}
\usage{
f_joint_root(f, start, tol = 1e-08, n.max = 1000)
}
\arguments{
\item{f}{function: A function that receives a vector and return a vector of the same size and a matrix representing its derivatives.}

\item{start}{vector: The initial value to start the algorithm.}

\item{tol}{numeric: The tolerance for the solution.}

\item{n.max}{numeric: The maximum number of iterations allowed.}
}
\value{
A list containing:
\itemize{
   \item root vector: The solution for the system f(x)=0.
   \item f.root vector: The function f evaluated at the root.
   \item iter numeric: The number of steps taken.
}
}
\description{
Calculates the root of a function given an initial value and a function to calculate its derivatives.
}
\keyword{internal}
