% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{f_root}
\alias{f_root}
\title{f_root}
\usage{
f_root(f, df, start, tol = 1e-08, n_max = 1000)
}
\arguments{
\item{f}{function: A function that receives a vector and return a vector of the same size.}

\item{df}{function: A function that receives a vector and return the derivatives of f in relation to it's arguments (must return a matrix, if f returns a vector).}

\item{start}{vector: The initial value to start the algorithm.}

\item{tol}{numeric: The tolerance for the solution error.}

\item{n_max}{numeric: The maximum number of iterations allowed.}
}
\value{
A list containing:
\itemize{
   \item root vector: The solution for the system.
   \item f.root vector: The function f evaluated at the root.
   \item iter numeric: The number of steps taken.
}
}
\description{
Calculates the root of a function given an initial value and a function to calculate it's derivatives.
}
\keyword{internal}
