% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{Fgamma_pred}
\alias{Fgamma_pred}
\title{Fgamma_pred}
\usage{
Fgamma_pred(conj_param, outcome = NULL, parms = list(), pred_cred = 0.95)
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the distribution of the linear predictor.}

\item{outcome}{Vector or matrix (optional): The observed values at the current time. Not used in this function.}

\item{parms}{List: A list of extra parameters for the model. For this function, it must contain the shape parameter phi of the observational model.}

\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector/matrix: the mean of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector/matrix: the variance of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector/matrix: the percentile of 100*((1-pred_cred)/2)% of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector/matrix: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item log.like vector: the The log likelihood for the outcome given the conjugated parameters.
}
}
\description{
Calculate the values for the predictive distribution given the values of the parameter of the distribution of the linear predictor.
The data is assumed to have Gamma distribution with unknown shape phi and unknown mean having log-Normal distribution.
In this scenario, the marginal distribution of the data is obtained via Monte Carlo.
}
\seealso{
Other {auxiliary functions for a Gamma outcome with unknowned shape}: 
\code{\link{Fgamma_pred_alt}()},
\code{\link{convert_FGamma_Normal}()},
\code{\link{convert_Normal_FGamma}()},
\code{\link{update_FGamma_alt}()},
\code{\link{update_FGamma}()}
}
\concept{{auxiliary functions for a Gamma outcome with unknowned shape}}
\keyword{internal}
