% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{Fgamma_pred_alt}
\alias{Fgamma_pred_alt}
\title{Fgamma_pred_alt}
\usage{
Fgamma_pred_alt(conj.param, outcome = NULL, parms = list(), pred.cred = 0.95)
}
\arguments{
\item{conj.param}{list or data.frame: The parameters of the distribution of the linear predictor.}

\item{outcome}{numeric or matrix (optional): The observed values at the current time. Not used in this function.}

\item{parms}{list: A list of extra parameters for the model. For this function, it must contain the shape parameter phi of the observational model.}

\item{pred.cred}{numeric: the desired credibility for the credibility interval.}
}
\value{
A list containing the following values:
\itemize{
   \item pred numeric/matrix: the mean of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item var.pred numeric/matrix: the variance of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item icl.pred numeric/matrix: the percentile of 100*((1-pred.cred)/2)% of the predictive distribution of a next observation. Same type and shape as the parameter in model.
   \item icu.pred numeric/matrix: the percentile of 100*(1-(1-pred.cred)/2)% of the predictive distribution of a next observation. Same type and shape as the parameter in model.
}
}
\description{
Calculate the values for the predictive distribution given the values of the parameter of the distribution of the linear predictor.
The data is assumed to have Gamma distribution with unknown shape and mean parameters having log-Normal distribution.
In this scenario, the marginal distribution of the data is obtained via Monte Carlo.
}
\seealso{
Other {auxiliary functions for a Gamma outcome with unknowned shape}: 
\code{\link{update_FGamma_alt}()}
}
\concept{{auxiliary functions for a Gamma outcome with unknowned shape}}
\keyword{internal}
