% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.values.fitted_dlm}
\alias{fitted.values.fitted_dlm}
\title{fitted.values.fitted_dlm}
\usage{
\method{fitted}{values.fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to eval_past}
}
\description{
fitted.values method for class fitted_dlm
}
\keyword{internal}
