% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{forecast.fitted_dlm}
\alias{forecast.fitted_dlm}
\title{forecast.fitted_dlm}
\usage{
\method{forecast}{fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to forecast_dlm}
}
\description{
forecast method for class fitted_dlm.
}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{coef.fitted_dlm}()},
\code{\link{coefficients.fitted_dlm}()},
\code{\link{dlm_sampling}()},
\code{\link{eval_past}()},
\code{\link{forecast_dlm}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
\keyword{internal}
