% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{forecast}
\alias{forecast}
\title{forecast}
\usage{
forecast(
  model,
  t = 1,
  outcome = NULL,
  offset = NULL,
  FF = NULL,
  G = NULL,
  D = NULL,
  W = NULL,
  plot = ifelse(requireNamespace("plotly", quietly = TRUE), "plotly", TRUE),
  pred_cred = 0.95
)
}
\arguments{
\item{model}{<undefined class> or list: The fitted model to be use for predictions.}

\item{t}{Numeric: Time window for prediction.}

\item{outcome}{List (optional): A named list containing the observed values in the prediction window. Note that the names in the list must be the same as the names passed during the fitting process.}

\item{offset}{Matrix or scalar: offset for predictions. Should have dimensions k x t, where k is the number of outcomes of the model. If offset is not specified, the last value observed by the model will be used.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x k x t, where n is the number of latent variables, k is the number of outcomes in the model. If not specified, the last value given to the model will be used.}

\item{G}{Array: A 3D-array containing the evolution matrix for each time. It's dimension should be n x n x t, where n is the number of latent variables. If not specified, the last value given to the model will be used.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x t, where n is the number of latent variables and T is the time series length. If not specified, the last value given to the model will be used in the first step, and 1 will be use thereafter.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be n x n x t, where n is the number of latent variables and T is the time series length. If not specified, 0 will be used.}

\item{plot}{Bool: A flag indicating if a plot should be produced.}

\item{pred_cred}{Numeric: The credibility level for the I.C. intervals.}
}
\value{
A list containing:
\itemize{
   \item pred Matrix: A matrix with the predictive mean at each time. Dimensions are k x t, where k is the number of outcomes.
   \item var.pred Array: A 3D-array with the predictive covariance matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item icl.pred Matrix: A matrix with the lower bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item icu.pred Matrix: A matrix with the upper bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item at Matrix: A matrix with the values for the linear predictor at each time. Dimensions are k x t, where k is the number of outcomes.
   \item Rt Array: A 3D-array with the covariance of the linear predictor matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item plot (if so chosen): A plotly or ggplot object.
}
}
\description{
Makes predictions for t times ahead using the fitted model.
}
\examples{
T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)

forecast(fitted_data, 20)$plot

}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{dlm_sampling}()},
\code{\link{eval_past}()},
\code{\link{fit_model}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
