% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{formula.to.structure}
\alias{formula.to.structure}
\title{formula.to.structure}
\usage{
\method{formula}{to.structure}(formula, data, label = "mu")
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which glm is called.}

\item{label}{An optional character naming the linear predictor.}
}
\description{
formula.to.structure
}
\keyword{internal}
