% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gastroBR}
\alias{gastroBR}
\title{Hospital admissions from gastroenteritis in Brazil}
\format{
A data frame with 4212 rows and 4 variables:
\describe{
  \item{UF}{The abbreviated state name.}
  \item{Date}{The date of the observation. Note that the day is only a placeholder and is just a placeholder.}
  \item{Admissions}{The number hospital admissions.}
  \item{Population}{The estimated population.}
}
}
\source{
Admissions: \url{https://datasus.saude.gov.br/informacoes-de-saude-tabnet/}

Population: \url{https://www.ibge.gov.br/estatisticas/sociais/populacao.html}
}
\usage{
gastroBR
}
\description{
A dataset containing the number of Hospital admissions from gastroenteritis in Brazil, per state, from 2010 to 2022 by month.
}
\keyword{datasets}
