% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{generic_smoother}
\alias{generic_smoother}
\title{generic_smoother}
\usage{
generic_smoother(mt, Ct, at, Rt, G, G_labs)
}
\arguments{
\item{mt}{Matrix: A matrix containing the filtered mean of the latent variables at each time. Each row should represent one variable.}

\item{Ct}{Array: A 3D-array representing the filtered covariance matrix of the latent variables at each time. The third dimension should represent the time index.}

\item{at}{Matrix: A matrix containing the one-step-ahead mean of the latent variables at each time based upon the filtered mean. Each row should represent one variable.}

\item{Rt}{Array: A 3D-array representing the one-step-ahead covariance matrix of the latent variables at each time based upon the filtered covariance matrix. The third dimension should represent the time index.}

\item{G}{Array: A 3D-array representing the transition matrix of the model at each time.}

\item{G_labs}{Matrix: A character matrix containing the type associated with each value in G.}
}
\value{
List: The smoothed mean (mts) and covariance (Cts) of the latent variables at each time. Their dimension follows, respectively, the dimensions of mt and Ct.
}
\description{
Generic smoother for all models.
}
\details{
For the models covered in this package, we always assume that the latent states have Multivariate Normal distribution. With that assumption, we can use Kalman Smoother algorithm to calculate the posterior of the states at each time, given everything that has been observed (assuming that we already know the filtered distribution of the states).

For the details about the implementation see \insertCite{ArtigoPacote;textual}{kDGLM}.

For the details about the algorithm implemented see \insertCite{ArtigokParametrico;textual}{kDGLM}, \insertCite{Petris-DLM;textual}{kDGLM}, chapter 2, \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 4, and \insertCite{Kalman_filter_origins;textual}{kDGLM}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_model}}

\code{\link{analytic_filter}}
}
\keyword{internal}
