% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{ginv}
\alias{ginv}
\title{ginv}
\usage{
ginv(S)
}
\arguments{
\item{S}{A covariance matrix}
}
\description{
This function receives a covariance matrix S and calculates the generalized inverse of S.
}
\keyword{internal}
