% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{harmonic_block}
\alias{harmonic_block}
\title{Structural blocks for seasonal trends and regressions}
\usage{
harmonic_block(
  ...,
  period,
  name = "Var_Sazo",
  D = 1,
  W = 0,
  m0 = 0,
  C0 = c(NA, 1)
)
}
\arguments{
\item{...}{Named values for the planing matrix.}

\item{period}{Positive integer: The size of the harmonic cycle.}

\item{name}{String: An optional argument providing the name for this block. Can be useful to identify the models with meaningful labels, also, the name used will be used in some auxiliary functions.}

\item{D}{Array, Matrix, vector or  scalar: The values for the discount factors at each time. If D is a array, it's dimensions should be n x n x t, where n is the order of the polynomial block and t is the length of the outcomes. If D is a matrix, it's dimensions should be n x n and it's values will be used for each time. If D is a vector or scalar, a discount factor matrix will be created as a diagonal matrix with the values of D in the diagonal.}

\item{W}{Array, Matrix, vector or  scalar: The values for the covariance matrix for the noise factor at each time. If W is a array, it's dimensions should be n x n x t, where n is the order of the polynomial block and t is the length of the outcomes. If W is a matrix, it's dimensions should be n x n and it's values will be used for each time. If W is a vector or scalar, a discount factor matrix will be created as a diagonal matrix with the values of W in the diagonal.}

\item{m0}{Vector or scalar: The prior mean for the latent variables associated with this block. If m0 is a vector, it's dimension should be equal to the order of the polynomial block. If m0 is a scalar, it's value will be used for all latent variables.}

\item{C0}{Matrix, vector or scalar: The prior covariance matrix for the latent variables associated with this block. If C0 is a matrix, it's dimensions should be n x n. If W is a vector or scalar, a covariance matrix will be created as a diagonal matrix with the values of C0 in the diagonal.}
}
\value{
An object of the class dlm_block containing the following values:
\itemize{
   \item FF Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x m x T, where n is the number of latent variables, m is the number of outcomes in the model and T is the time series length.
   \item G Matrix: The state evolution matrix.
   \item D Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x T, where n is the number of latent variables and T is the time series length.
   \item W Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be the same as D.
   \item m0 Vector: The prior mean for the latent vector.
   \item C0 Matrix: The prior covariance matrix for the latent vector.
   \item names list: A list containing the variables indexes by their name.
   \item period Positive integer: Same as argument.
   \item n Positive integer: The number of latent variables associated with this block (2).
   \item t Positive integer: The number of time steps associated with this block. If 1, the block is compatible with blocks of any time length, but if t is greater than 1, this block can only be used with blocks of the same time length.
   \item k Positive integer: The number of outcomes associated with this block. This block can only be used with blocks with the same outcome length.
   \item var_names Vector: The name of the linear predictors associated with this block,
   \item type Character: The type of block (Harmonic).
}
}
\description{
Creates the structure for a harmonic block with desired periodicity.
}
\details{
For the ..., D, W, m0 and C0 arguments, the user may set one or more of it's values as a string.
By doing so, the user will leave the block partially undefined and it can no longer be used in the \code{\link{fit_model}} function.
Instead, the user must use the \code{\link{search_model}} function to search the best hyper parameters among a defined range of possible values.
See the \code{\link{search_model}} function for details on it's usage.

For the details about the implementation see \insertCite{ArtigoPacote;textual}{kDGLM}.

For the details about the modelling of seasonal trends using harmonics in the context of DLM's, see \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 8.

For the details about dynamic regression models in the context of DLM's, see \insertCite{WestHarr-DLM;textual}{kDGLM}, chapters 6 and 9.
}
\examples{
# Creating seasonal structure for a model with 2 outcomes.
# One block is created for each outcome
# with each block being associated with only one of the outcomes.
season_1 <- harmonic_block(alpha1 = 1, period = 3)
season_2 <- harmonic_block(alpha2 = 1, period = 6)

# Creating a block with shared effect between the outcomes
season_3 <- harmonic_block(alpha = 1, alpha2 = 1, period = 12)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_model}}

\code{\link{search_model}}

Other {auxiliary functions for structural blocks}: 
\code{\link{AR_block}()},
\code{\link{block_merge}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{correlation_block}()},
\code{\link{polynomial_block}()}
}
\concept{{auxiliary functions for structural blocks}}
