% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{if.na}
\alias{if.na}
\title{if.na}
\usage{
if.na(vec, val)
}
\arguments{
\item{vec}{A vector or matrix.}

\item{val}{The value to replace NA with.}
}
\description{
This function is wrapper for ifelse(is.null(.),.,.)
}
\keyword{internal}
