% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{if.null}
\alias{if.null}
\title{if.null}
\usage{
if.null(vec, val)
}
\arguments{
\item{vec}{A vector or matrix.}

\item{val}{The value to replace NULL with.}
}
\description{
This function is wrapper for ifelse(is.null(.),.,.)
}
\keyword{internal}
