% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{lcm}
\alias{lcm}
\title{lcm}
\usage{
lcm(x)
}
\arguments{
\item{x}{numeric: A sequence of integers.}
}
\value{
The least common multiple.
}
\description{
Calculates the least common multiple of a set of integer. Internal use only.
}
\keyword{internal}
