% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multnom.R
\name{Multinom}
\alias{Multinom}
\title{Multinom outcome for kDGLM models}
\usage{
Multinom(p, outcome, offset = outcome^0, alt_method = FALSE)
}
\arguments{
\item{p}{character: a vector with the name of the linear predictor associated with the probality of each category (except the base one, which is assumed to be the last).}

\item{outcome}{vector: Values of the observed data.}

\item{offset}{vector: The offset at each observation. Must have the same shape as outcome.}
}
\value{
A object of the class dlm_distr
}
\description{
Creates an outcome with Multinomial distribution with the chosen parameters.
}
\details{
For evaluating the posterior parameters, we use the method proposed in \insertCite{ArtigokParametrico;textual}{kDGLM}.

For the details about the implementation see  \insertCite{ArtigoPacote;textual}{kDGLM}.
}
\examples{

# Multinomial case
T <- 200
y1 <- rpois(T, exp(5 + (-T:T / T) * 5))
y2 <- rpois(T, exp(6 + (-T:T / T) * 5 + sin((-T:T) * (2 * pi / 12))))
y3 <- rpois(T, exp(5))

y <- cbind(y1, y2, y3)

level1 <- polynomial_block(p1 = 1, order = 2)
level2 <- polynomial_block(p2 = 1, order = 2)
season <- harmonic_block(p2 = 1, period = 12)
outcome <- Multinom(p = c("p1", "p2"), outcome = y)

fitted_data <- fit_model(level1, level2, season, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, smooth = TRUE)$plot

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_model}}

Other {auxiliary functions for a creating outcomes}: 
\code{\link{Gamma}()},
\code{\link{Normal}()},
\code{\link{Poisson}()},
\code{\link{report_distr}()}
}
\concept{{auxiliary functions for a creating outcomes}}
