% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noticeSARI}
\alias{noticeSARI}
\title{SARI data from Belo Horizonte}
\format{
A data frame with 65404 rows and 7 variables:
\describe{
  \item{ref.week}{The reference week, counting since the monitoring begun.}
  \item{reported.k.week}{The number of cases occurred in the period and reported until the k weeks after the reference week.}
  \item{occured}{The total number of cases reported (at any time).}
  ...
}
}
\source{
\url{https://opendatasus.saude.gov.br/}
}
\usage{
noticeSARI
}
\description{
A dataset containing reports from Severe Acute Respiratory Illness (SARI) from 2020 to April 2022 by week.
}
\keyword{datasets}
