% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dlm_block}
\alias{print.dlm_block}
\title{print.dlm_block}
\usage{
\method{print}{dlm_block}(x, ...)
}
\arguments{
\item{x}{A dlm_block object.}

\item{...}{Arguments passed to summary.dlm_block}
}
\description{
print.dlm_block
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{*.dlm_block}()},
\code{\link{+.dlm_block}()},
\code{\link{AR_block}()},
\code{\link{PCAR_block}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{regression_block}()},
\code{\link{report_block}()},
\code{\link{summary.dlm_block}()}
}
\concept{{auxiliary functions for structural blocks}}
\keyword{internal}
