% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dlm_distr}
\alias{print.dlm_distr}
\title{print.dlm_distr}
\usage{
\method{print}{dlm_distr}(x, ...)
}
\arguments{
\item{x}{A dlm_distr object.}

\item{...}{Arguments passed to summary.dlm_distr}
}
\description{
print.dlm_distr
}
\seealso{
Other {auxiliary visualization functions for the fitted.dlm class}: 
\code{\link{print.searched_dlm}()},
\code{\link{summary.fitted_dlm}()},
\code{\link{summary.searched_dlm}()}
}
\concept{{auxiliary visualization functions for the fitted.dlm class}}
\keyword{internal}
