% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.searched_dlm}
\alias{print.searched_dlm}
\title{print.searched_dlm}
\usage{
\method{print}{searched_dlm}(x, ...)
}
\arguments{
\item{x}{A searched_dlm object.}

\item{...}{Arguments passed to summary.searched_dlm}
}
\description{
print.searched_dlm
}
\keyword{internal}
