% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{specify}
\alias{simulate}
\alias{forecast}
\alias{fit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}, \code{\link[generics]{forecast}}, \code{\link[generics]{specify}}}

  \item{stats}{\code{\link[stats]{simulate}}}
}}

