% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{report_searched_dlm}
\alias{report_searched_dlm}
\title{Summary for a searched_dlm object}
\usage{
report_searched_dlm(searched_dlm)
}
\arguments{
\item{searched_dlm}{A searched_dlm object.}
}
\description{
Prints a report for a searched_dlm object.
}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lat_var}()},
\code{\link{plot_lin_pred}()},
\code{\link{report_dlm}()},
\code{\link{show_fit}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
\keyword{internal}
