% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{rmvnorm}
\alias{rmvnorm}
\title{rmvnorm}
\usage{
rmvnorm(n, mu, Sigma, norm.x = matrnorm(k, n))
}
\arguments{
\item{n}{integer: The sample size.}

\item{mu}{Vector: The mean vector}

\item{Sigma}{Matrix: The Covariance matrix.}
}
\description{
Obtains a sample from a multivariate normal distribution.
}
\keyword{internal}
