% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smoothing}
\alias{smoothing}
\title{Auxiliary function for model smoothing}
\usage{
smoothing(model)
}
\arguments{
\item{model}{A fitted_dlm object.}
}
\value{
A fitted_dlm object with smoothed means (mts) and covariance matrix (Cts) for each observation.
}
\description{
Auxiliary function for model smoothing
}
