% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary.dlm_block}
\alias{summary.dlm_block}
\title{Summary for a kDGLM structure}
\usage{
\method{summary}{dlm_block}(object, ...)
}
\arguments{
\item{object}{A dlm_block object.}
}
\description{
Prints a report for a dlm_block object.
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{AR_block}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{intervention}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{regression_block}()},
\code{\link{set.block.value}()}
}
\concept{{auxiliary functions for structural blocks}}
\keyword{internal}
