% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dlm_distr}
\alias{summary.dlm_distr}
\title{summary.dlm_distr}
\usage{
\method{summary}{dlm_distr}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Not used.}
}
\description{
summary method for class dlm_distr.
}
\seealso{
Other {auxiliary functions for a creating outcomes}: 
\code{\link{Gamma}()},
\code{\link{Multinom}()},
\code{\link{Normal}()},
\code{\link{Poisson}()},
\code{\link{report_distr}()}

Other {Reports for dlm_distr objects.}: 
\code{\link{report_distr}()}
}
\concept{{Reports for dlm_distr objects.}}
\concept{{auxiliary functions for a creating outcomes}}
\keyword{internal}
