% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.fitted_dlm}
\alias{summary.fitted_dlm}
\title{summary.fitted_dlm}
\usage{
\method{summary}{fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to report_dlm}
}
\description{
summary method for class fitted_dlm.
}
\keyword{internal}
