% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.searched_dlm}
\alias{summary.searched_dlm}
\title{summary.searched_dlm}
\usage{
\method{summary}{searched_dlm}(object, ...)
}
\arguments{
\item{object}{A searched_dlm object.}

\item{...}{Arguments passed to report_searched_dlm}
}
\description{
summary method for class searched_dlm
}
\keyword{internal}
