% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{system_full_gamma}
\alias{system_full_gamma}
\title{system_full_gamma}
\usage{
system_full_gamma(x, parms)
}
\arguments{
\item{x}{vector: current tau values.}

\item{parms}{list: auxiliary values for the system.}
}
\value{
A vector with the values of the system \insertCite{@see ArtigokParametrico;textual}{kDGLM}.
}
\description{
Evaluate the compatibilizing equation for the full gamma model \insertCite{@see ArtigokParametrico;textual}{kDGLM}.
}
\keyword{internal}
