% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update.fitted_dlm}
\alias{update.fitted_dlm}
\title{update.fitted_dlm}
\usage{
\method{update}{fitted_dlm}(model, ...)
}
\arguments{
\item{model}{fitted_dlm: The fitted model to be updated.}

\item{...}{Extra variables necessary for updating (covariates, observed values, etc.).}
}
\value{
A fitted_dlm object.
}
\description{
Updates a fitted model.
}
\examples{

level <- polynomial_block(rate = 1, order = 2, D = 0.95)
season <- harmonic_block(rate = 1, order = 2, period = 12, D = 0.975)

# Only first 100 observations (for the sake of the example)
outcome <- Poisson(lambda = "rate", data = c(AirPassager)[1:100])

fitted.data <- fit_model(level, season,
  AirPassengers = outcome
)

updated.fit <- update(fitted.data, AirPassager = list(data = c(AirPassager)[101:144]))
# If a offset was present, the user should pass its value when updating
# updated.fit=update(fitted.data,
#                     AirPassager=list(
#                      data=c(AirPassager)[101:144],
#                      offset= ... ))

}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{coef.fitted_dlm}()},
\code{\link{coefficients.fitted_dlm}()},
\code{\link{dlm_sampling}()},
\code{\link{forecast.fitted_dlm}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
