% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{update_Gamma}
\alias{update_Gamma}
\title{update_Gamma}
\usage{
update_Gamma(conj_prior, ft, Qt, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Inverse-Gamma (alpha,beta).}

\item{ft}{vector: A vector representing the means from the normal distribution. Not used in the default method.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution. Not used in the default method.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. For this kernel, parms should containg the shape parameter (phi) for the observational gamma model.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Inverse-Gamma, assuming that the observed values came from a Gamma model from which the shape parameter (phi) is known and the mean (mu) have prior distribution Inverse-Gamma.
}
\seealso{
Other {auxiliary functions for a Gamma outcome with known shape}: 
\code{\link{convert_Gamma_Normal}()},
\code{\link{convert_Normal_Gamma}()},
\code{\link{gamma_pred}()},
\code{\link{update_Gamma_alt}()}
}
\concept{{auxiliary functions for a Gamma outcome with known shape}}
\keyword{internal}
