% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multnom.R
\name{update_Multinom}
\alias{update_Multinom}
\title{update_Multinom}
\usage{
update_Multinom(conj.param, ft, Qt, y, parms = list())
}
\arguments{
\item{conj.param}{list: A vector containing the concentration parameters of the Dirichlet.}

\item{ft}{vector: A vector representing the means from the normal distribution. Not used in the default method.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution. Not used in the default method.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this kernel.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Dirichlet, assuming that the observed values came from a Multinomial model from which the number of trials is known and the prior distribution for the probabilities of each category have joint distribution Dirichlet.
}
\seealso{
Other auxiliary functions for a Multinomial outcome: 
\code{\link{convert_Multinom_Normal}()},
\code{\link{convert_Normal_Multinom}()},
\code{\link{multnom_pred}()}
}
\concept{auxiliary functions for a Multinomial outcome}
\keyword{internal}
