% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{update_NG2}
\alias{update_NG2}
\title{update_NG}
\usage{
update_NG2(conj.param, ft, Qt, y, parms = list())
}
\arguments{
\item{conj.param}{list: A vector containing the parameters of the Normal-Gamma (mu0,c0,alpha,beta).}

\item{ft}{numeric: A vector representing the means from the normal distribution. Not used in the default method.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution. Not used in the default method.}

\item{y}{numeric: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this kernel.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Normal-Gamma, assuming that the observed values came from a Normal model from which the prior distribution for the mean and the precision have joint distribution Normal-Gamma
}
\keyword{internal}
