% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{update_NG_alt2}
\alias{update_NG_alt2}
\title{update_NG_alt}
\usage{
update_NG_alt2(conj.param, ft, Qt, y, parms = list())
}
\arguments{
\item{conj.param}{list: A vector containing the parameters of the Normal-Gamma (mu0,c0,alpha,beta). Not used in the alternative method.}

\item{ft}{numeric: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{y}{numeric: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this kernel.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate the (approximated) posterior parameter for the linear predictors, assuming that the observed values came from a Normal model from which the mean and the log precision have prior distribution in the Normal family.
}
\details{
For evaluating the posterior parameters, we use a modified version of the method proposed in \insertCite{ArtigokParametrico;textual}{kDGLM}.

For the details about the implementation see  \insertCite{ArtigoPacote;textual}{kDGLM}.

For the detail about the modification of the method proposed in \insertCite{ArtigokParametrico;textual}{kDGLM}, see \insertCite{ArtigoAltMethod;textual}{kDGLM}.
}
\references{
\insertAllCited{}
}
\seealso{
Other {auxiliary functions for a Normal outcome}: 
\code{\link{convert_multi_NG_Normal}()},
\code{\link{multi_normal_gamma_pred}()},
\code{\link{normal_pred}()},
\code{\link{update_NG_alt}()},
\code{\link{update_Normal}()}
}
\concept{{auxiliary functions for a Normal outcome}}
\keyword{internal}
