% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_poisson.R
\name{update_Poisson}
\alias{update_Poisson}
\title{update_Poisson}
\usage{
update_Poisson(conj.param, ft, Qt, y, parms)
}
\arguments{
\item{conj.param}{list: A vector containing the parameters of the Gamma (alpha,beta).}

\item{ft}{numeric: A vector representing the means from the normal distribution. Not used in the default method.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution. Not used in the default method.}

\item{y}{numeric: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this kernel.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Gamma, assuming that the observed values came from a Poisson model from which the rate parameter (lambda) have prior distribution Gamma.
}
\seealso{
Other auxiliary functions for a Poisson outcome: 
\code{\link{convert_Normal_Poisson}()},
\code{\link{convert_Poisson_Normal}()},
\code{\link{poisson_pred}()}
}
\concept{auxiliary functions for a Poisson outcome}
\keyword{internal}
