% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{var_decomp}
\alias{var_decomp}
\title{var_decomp}
\usage{
var_decomp(S)
}
\arguments{
\item{S}{A covariance matrix}
}
\description{
This function receives a covariance matrix S and creates a matrix Q, so that t(Q) %*% Q=S.
}
\keyword{internal}
