% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{zero_sum_prior}
\alias{zero_sum_prior}
\title{Zero sum prior}
\usage{
zero_sum_prior(block, var.index = 1:block$n)
}
\arguments{
\item{block}{dlm_block: The structural block.}

\item{var.index}{Integer: The index of the variables from which to set the prior.}
}
\value{
A dlm_block object with the desired prior.
}
\description{
Set the prior of a structural block to be such that the latent variables sum zero.
The covariance matrix of the evolution and the drift parameter are also altered to guarantee that the zero sum condition will always hold.
}
\details{
For the details about the implementation see \insertCite{ArtigoPacote;textual}{kDGLM}.
}
\references{
\insertAllCited{}
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{*.dlm_block}()},
\code{\link{+.dlm_block}()},
\code{\link{AR_block}()},
\code{\link{CAR_prior}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{print.dlm_block}()},
\code{\link{regression_block}()},
\code{\link{summary.dlm_block}()}
}
\concept{{auxiliary functions for structural blocks}}
