% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{AR_block}
\alias{AR_block}
\title{Structural blocks for auto regressive trends and regressions}
\usage{
AR_block(
  ...,
  order,
  noise_var,
  noise_discount = 1,
  pulse = 0,
  name = "Var_AR",
  AR_support = "free",
  D = 1,
  h = 0,
  H = 0,
  a1 = c(0, rep(0, order - 1)),
  R1 = 0.25,
  monitoring = rep(FALSE, order),
  a1_states = 0,
  R1_states = 9,
  h_states = 0,
  monitoring_states = TRUE,
  a1_pulse = 0,
  R1_pulse = 9,
  D_pulse = 1,
  h_pulse = 0,
  H_pulse = 0,
  monitoring_pulse = NA
)
}
\arguments{
\item{...}{Named values for the planning matrix.}

\item{order}{Positive integer: The order of the AR block.}

\item{noise_var}{Non-negative scalar: The variance of the white noise added to the latent state.}

\item{noise_discount}{Vector or scalar: The value for the discount factor associated with the current latent state. If noise_discount is a vector, it should have size t and it is interpreted as the discount factor at each observed time. If D is a scalar, the same discount will be used for all observation.}

\item{pulse}{Vector or scalar: An optional argument providing the values for the pulse for a Transfer Function. Default is 0 (no Transfer Function).}

\item{name}{String: An optional argument providing the name for this block. Can be useful to identify the models with meaningful labels, also, the name used will be used in some auxiliary functions.}

\item{AR_support}{String: Either "constrained" or "free" (default). If AR_support is "constrained", then the AR coefficients will be forced to be on the interval (-1,1), otherwise, the coefficients will be unrestricted. Beware that, under no restriction on the coefficients, there is no guarantee that the estimated coefficients will imply in a stationary process, furthermore, if the order of the AR block is greater than 1. As such the restriction of the coefficients support is only available for AR blocks with order equal to 1.}

\item{D}{Array, Matrix, vector or scalar: The values for the discount factors associated with the AR coefficients at each time. If D is an array, it's dimensions should be n x n x t, where n is the order of the AR block and t is the length of the outcomes. If D is a matrix, it's dimensions should be n x n and the same discount matrix will be used in all observations. If D is a vector, it should have size t and it is interpreted as the discount factor at each observed time (same discount for all variable). If D is a scalar, the same discount will be used for all AR coefficients at all times.}

\item{h}{Matrix, vector or scalar: A drift to be add in the AR coefficients after the temporal evolution (can be interpreted as the mean of the random noise at each time). If a matrix, it's dimension should be n x T, where n is the order of the AR block and T is the length of the series. If a scalar, the passed value will be used for all coefficients at each time.}

\item{H}{Array, Matrix, vector or scalar: The values for the covariance matrix for the noise factor associated with the AR coefficients at each time. If H is a array, it's dimensions should be n x n x t, where n is the order of the AR block and t is the length of the outcomes. If H is a matrix, it's dimensions should be n x n and its values will be used for each time. If H is a vector or scalar, a discount factor matrix will be created as a diagonal matrix with the values of H in the diagonal.}

\item{a1}{Vector or scalar: The prior mean for the AR coefficients associated with this block at time 1. If a1 is a vector, it's dimension should be equal to the order of the AR block. If a1 is a scalar, it's value will be used for all coefficients. If the coefficients are restricted to the interval (-1,1), the a1 is interpreted as the mean for atanh(rho), where rho is the AR coefficient.}

\item{R1}{Matrix, vector or scalar: The prior covariance matrix for the coefficients associated with this block at time 1. If R1 is a matrix, it's dimensions should be n x n, where n is the order of the AR block. If R1 is a vector or scalar, a covariance matrix will be created as a diagonal matrix with the values of R1 in the diagonal. If the coefficients are restricted to the interval (-1,1), the R1 is interpreted as the covariance matrix for atanh(rho), where rho is the AR coefficient.}

\item{monitoring}{Vector: A vector of flags indicating which AR coefficients should be monitored (if automated monitoring is used). Its size should be n, where n is the order of the AR block. The default is that no coefficient should be monitored.}

\item{a1_states}{Vector or scalar: The prior mean for the states associated with this block at time 1. If a1_states is a vector, it's dimension should be equal to the order of the AR block. If a1_states is a scalar, it's value will be used for all coefficients.}

\item{R1_states}{Matrix, vector or scalar: The prior covariance matrix for the states associated with this block at time 1. If R1_states is a matrix, it's dimensions should be n x n, where n is the order of the AR block. If R1_state is a vector or scalar, a covariance matrix will be created as a diagonal matrix with the values of R1_state in the diagonal.}

\item{h_states}{Vector or scalar: A drift to be add in the states after the temporal evolution (can be interpreted as the mean of the random noise at each time). If a vector, it should have size t, and each value will be applied in their respective time. If a scalar, the passed value will be used for all observations.}

\item{monitoring_states}{bool: A flag indicating if the latent state should be monitored (if automated monitoring is used). The default is TRUE.}

\item{a1_pulse}{Vector or scalar: The prior mean for the coefficients associated with the pulses at time 1. If a1_pulse is a vector, it's dimension should be equal to the number of pulses. If a1_pulse is a scalar, it's value will be used for all coefficients.}

\item{R1_pulse}{Matrix, vector or scalar: The prior covariance matrix for the coefficients associated with the pulses at time 1. If R1_pulse is a matrix, it's dimensions should be n x n, where n is the number of pulses. If R1_pulse is a vector or scalar, a covariance matrix will be created as a diagonal matrix with the values of R1_pulse in the diagonal.}

\item{D_pulse}{Array, Matrix, vector or scalar: The values for the discount factors associated with the pulse coefficients at each time. If D_pulse is an array, it's dimensions should be n x n x t, where n is the number of pulses and t is the length of the outcomes. If D_pulse is a matrix, it's dimensions should be n x n and the same discount matrix will be used in all observations. If D_pulse is a vector, it should have size t and it is interpreted as the discount factor at each observed time (same discount for all variable). If D is a scalar, the same discount will be used for all pulse coefficients at all times.}

\item{h_pulse}{Matrix, vector or scalar: A drift to be add in the pulse effect after the temporal evolution (can be interpreted as the mean of the random noise at each time). If a matrix, it's dimension should be n x T, where n is the number of pulses and T is the length of the series. If a scalar, the passed value will be used for all latent variable at each time.}

\item{H_pulse}{Array, Matrix, vector or scalar: The values for the covariance matrix for the noise factor associated with pulse coefficients at each time. If H_pulse is an array, it's dimensions should be n x n x t, where n is the number of pulses and t is the length of the outcomes. If H_pulse is a matrix, it's dimensions should be n x n and its values will be used for each time. If H_pulse is a vector or scalar, a covariance matrix will be created as a diagonal matrix with the values of H_pulse in the diagonal.}

\item{monitoring_pulse}{Vector: A vector of flags indicating which pulse coefficients should be monitored (if automated monitoring is used). Its size should be n, where n is the number of pulses. The default is that no pulse coefficient should be monitored.}
}
\value{
A dlm_block object containing the following values:
\itemize{
   \item FF Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x m x T, where n is the number of latent variables, m is the number of outcomes in the model and T is the time series length.
   \item FF_labs Array: DESCRIPTION
   \item G Matrix: The state evolution matrix.
   \item D Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x T, where n is the number of latent variables and T is the time series length.
   \item H Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be the same as D.
   \item a1 Vector: The prior mean for the latent vector.
   \item R1 Matrix: The prior covariance matrix for the latent vector.
   \item var_names list: A list containing the variables indexes by their name.
   \item order Positive integer: Same as argument.
   \item n Positive integer: The number of latent variables associated with this block (2).
   \item t Positive integer: The number of time steps associated with this block. If 1, the block is compatible with blocks of any time length, but if t is greater than 1, this block can only be used with blocks of the same time length.
   \item k Positive integer: The number of outcomes associated with this block. This block can only be used with blocks with the same outcome length.
   \item pred_names Vector: The name of the linear predictors associated with this block.
   \item monitoring Vector: The combination of monitoring, monitoring_states and monitoring_pulse.
   \item type Character: The type of block (AR).
   \item AR_support Character: Same as argument.
}
}
\description{
Creates the structure for a Auto Regressive (AR) block (see \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 9) with desired order.
As the package suppose that the structure of the model is linear, a linearization is applied to the evolution equation, as described in \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 13.
This block also supports Transfer Functions, being necessary to specify the associated pulse when calling the AR_block function (see arg.).
}
\details{
For the ..., noise_var, D, H, a1, R1, a1_states, R1_states, D_states, a1_pulse, R1_pulse, D_pulse, S_pulse arguments, the user may set one or more of it's values as a string.
By doing so, the user will leave the block partially undefined and it can no longer be used in the \code{\link{fit_model}} function.
Instead, the user must use the \code{\link{search_model}} function to search the best hyper parameters among a defined range of possible values.
See the \code{\link{search_model}} function for details on it's usage.

For the details about the implementation see \insertCite{ArtigoPacote;textual}{kDGLM}.

For the details about Auto regressive models in the context of DLM's, see \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 9.

For the details about the linearization of non-linear evolution equations in the context of DLM's, see \insertCite{WestHarr-DLM;textual}{kDGLM}, chapter 13.

For the details about dynamic regression models in the context of DLM's, see \insertCite{WestHarr-DLM;textual}{kDGLM}, chapters 6 and 9.
}
\examples{

AR_block(mu = 1, pulse = rnorm(200), order = 3, noise_var = 0.1)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_model}}

Other {auxiliary functions for structural blocks}: 
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{polynomial_block}()},
\code{\link{regression_block}()},
\code{\link{report_block}()}
}
\concept{{auxiliary functions for structural blocks}}
