% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.fitted_dlm}
\alias{coef.fitted_dlm}
\title{coef.fitted_dlm}
\usage{
\method{coef}{fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to eval_past.}
}
\description{
coef method for class fitted_dlm
}
\keyword{internal}
