% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coef.fitted_dlm}
\alias{coef.fitted_dlm}
\title{coef.fitted_dlm}
\usage{
\method{coef}{fitted_dlm}(
  model,
  eval_t = seq_len(model$t),
  lag = -1,
  pred.cred = 0.95,
  eval.pred = TRUE
)
}
\arguments{
\item{model}{fitted_dlm: The fitted model to be use for evaluation.}

\item{eval_t}{vector: A vector of positive integers indicating the time index from which to extract predictions. Default is to extract all observed times in model.}

\item{lag}{positive integer: The relative offset for forecast. Values for time t will be calculated based on the filtered values of time t-h. If lag is negative, then the smoothed distribution will be used.}

\item{pred.cred}{Numeric: The credibility level for the I.C. intervals.}

\item{eval.pred}{Bool: A flag indicating if the predictions should be calculated.}
}
\value{
A list containing:
\itemize{
   \item data data.frame: A table with the model evaluated at each observed time.
   \item mt Matrix: The mean of the latent variables at each time. Dimensions are n x t, where t is the size of eval_t.
   \item Ct Array: A 3D-array containing the covariance matrix of the latent variable at each time. Dimensions are n x n x t.
   \item ft Matrix: The mean of the linear predictor at each time. Dimensions are k x t.
   \item Qt Array: A 3D-array containing the covariance matrix for the linear predictor at each time. Dimensions are k x k x t.
   \item log.like, mae, mase, rae, mse, interval.score: The metric value at each time.
   \item conj.param list: A list containing, for each outcome, a data.frame with the parameter of the conjugated distribution at each time.
}
}
\description{
Evaluates the predictive values for the observed values used to fit the model and its latent variables.
Predictions can be made with smoothed values or with filtered values with a time offset.
}
\examples{
# Poisson case
data <- c(AirPassengers)

level <- polynomial_block(rate = 1, order = 2, D = 0.95)
season <- harmonic_block(rate = 1, order = 2, period = 12, D = 0.975)

outcome <- Poisson(lambda = "rate", data = data)

fitted.data <- fit_model(level, season,
  AirPassengers = outcome
)

# var.vals <- coefficients(fitted.data)
# Or

}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{coefficients.fitted_dlm}()},
\code{\link{dlm_sampling}()},
\code{\link{forecast.fitted_dlm}()},
\code{\link{update.fitted_dlm}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
