% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coefficients.fitted_dlm}
\alias{coefficients.fitted_dlm}
\title{coefficients.fitted_dlm}
\usage{
\method{coefficients}{fitted_dlm}(object, ...)
}
\arguments{
\item{object}{A fitted_dlm object.}

\item{...}{Arguments passed to coef.}
}
\description{
coefficients method for class fitted_dlm
}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{coef.fitted_dlm}()},
\code{\link{dlm_sampling}()},
\code{\link{forecast.fitted_dlm}()},
\code{\link{update.fitted_dlm}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
\keyword{internal}
