% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{forecast.fitted_dlm}
\alias{forecast.fitted_dlm}
\title{forecast}
\usage{
\method{forecast}{fitted_dlm}(
  model,
  t = 1,
  plot = ifelse(requireNamespace("plotly", quietly = TRUE), "plotly",
    ifelse(requireNamespace("ggplot2", quietly = TRUE), "ggplot2", "base")),
  pred.cred = 0.95,
  ...
)
}
\arguments{
\item{model}{fitted_dlm: The fitted model to be use for predictions.}

\item{t}{Numeric: Time window for prediction.}

\item{plot}{Bool or String: A flag indicating if a plot should be produced. Should be one of FALSE, TRUE, 'base', 'ggplot2' or 'plotly'.}

\item{pred.cred}{Numeric: The credibility level for the I.C. intervals.}

\item{...}{Extra variables necessary for prediction (covariates, etc.).}
}
\value{
A list containing:
\itemize{
   \item data Data.frame: A data frame contain the mean, variance and credibility intervals for the outcomes, including both the observed data and the predictions for future observations.
   \item forecast Data.frame: Same as data, but restricted to predictions for future observations.
   \item outcomes List: A named list containing predictions for each outcome. Each element of this list is a list containing predictions (mean, variance and credibility intervals), the distribution of the linear predictor for the parameter of the observational model and the parameters of the predictive distribution (if available).
   \item mt Matrix: A matrix with the values for the latent states at each time. Dimensions are n x t, where n is the number of latent states
   \item Ct Array: A 3D-array with the covariance of the latent states at each time. Dimensions are n x n x t, where n is the number of linear predictors.
   \item ft Matrix: A matrix with the values for the linear predictors at each time. Dimensions are n x t, where n is the number of linear predictors
   \item Qt Array: A 3D-array with the covariance of the linear predictors at each time. Dimensions are n x n x t, where n is the number of linear predictors.
   \item plot (if so chosen): A plotly or ggplot object.
}
}
\description{
Makes predictions for t times ahead using the fitted model.
}
\examples{
data <- c(AirPassengers)

level <- polynomial_block(rate = 1, order = 2, D = 0.95)
season <- harmonic_block(rate = 1, order = 2, period = 12, D = 0.975)

outcome <- Poisson(lambda = "rate", data)

fitted.data <- fit_model(level, season,
  AirPassengers = outcome
)

forecast(fitted.data, 20)$plot

}
\seealso{
Other {auxiliary functions for fitted_dlm objects}: 
\code{\link{coef.fitted_dlm}()},
\code{\link{coefficients.fitted_dlm}()},
\code{\link{dlm_sampling}()},
\code{\link{update.fitted_dlm}()}
}
\concept{{auxiliary functions for fitted_dlm objects}}
