% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_helper.R
\name{if.nan}
\alias{if.nan}
\title{if.nan}
\usage{
if.nan(vec, val)
}
\arguments{
\item{vec}{A vector or matrix.}

\item{val}{The value to replace NaN with.}
}
\value{
A vector or matrix with the same dimensions as the input, where any NaN values have been replaced by the specified val argument.
}
\description{
This function is wrapper for ifelse(is.nan(vec),vec,val)
}
\keyword{internal}
