% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{Normal}
\alias{Normal}
\title{Normal outcome for kDGLM models}
\usage{
Normal(
  mu,
  V = NA,
  Tau = NA,
  Sigma = NA,
  Sd = NA,
  outcome,
  offset = outcome^0,
  alt_method = FALSE
)
}
\arguments{
\item{mu}{character: Name of the linear predictor associated with the mean parameter of the Normal distribution. The parameter is treated as unknown and equal to the associated linear predictor.}

\item{V}{character or numeric: If V is a character, it is interpreted as the names of the linear predictors associated with the variance parameter of the Normal distribution. If V is numeric, the variance is considered known and equal to the value of V, otherwise, the variance is considered unknown and equal to the exponential of the linear predictor informed in V. If the outcome is a Multivariate Normal, then V must be a matrix and, if the variance is unknown, the elements outside its main diagonal are treated as the linear predictor associated with the correlation between each coordinate of the outcome, otherwise V is treated as the covariance matrix. The user cannot specify V with Sigma, Tau or Sd.}

\item{Tau}{character or numeric: If Tau is a character, it is interpreted as the names of the linear predictors associated with the precisions parameter of the Normal distribution. If Tau is numeric, the precision is considered known and equal to the value of Tau, otherwise, the precision is considered unknown and equal to the exponential of the linear predictor informed in Tau. If the outcome is a Multivariate Normal, then Tau must be a matrix and, if the precision is unknown, the elements outside its main diagonal are treated as the linear predictor associated with the correlation between each coordinate of the outcome, otherwise Tau is treated as the precision matrix. The user cannot specify Tau with V, Sigma or Sd.}

\item{Sigma}{character or numeric: If Sigma is a character, it is interpreted as the names of the linear predictors associated with the variance parameter of the Normal distribution. If Sigma is numeric, the variance is considered known and equal to the value of Sigma, otherwise, the variance is considered unknown and equal to the exponential of the linear predictor informed in Sigma. If the outcome is a Multivariate Normal, then Sigma must be a matrix and, if the variance is unknown, the elements outside its main diagonal are treated as the linear predictor associated with the correlation between each coordinate of the outcome, otherwise Sigma is treated as the covariance matrix. The user cannot specify Sigma with V, Tau or Sd.}

\item{Sd}{character or numeric: If Sd is a character, it is interpreted as the names of the linear predictors associated with the standard deviation parameter of the Normal distribution. If Sd is numeric, the standard deviation is considered known and equal to the value of Sd, otherwise, the precision is considered unknown and equal to the exponential of the linear predictor informed by in Sd. If the outcome is a Multivariate Normal, then Tau must be a matrix and the elements outside its main diagonal are treated as the correlation (or the name of the linear predictor associated) between each coordinate of the outcome. The user cannot specify Sd with V, Tau or Sigma.}

\item{outcome}{vector: Values of the observed data.}

\item{offset}{vector: The offset at each observation. Must have the same shape as outcome.}
}
\value{
A object of the class dlm_distr
}
\description{
Creates an outcome with Normal distribution with the chosen parameters (can only specify 2,).
}
\details{
For evaluating the posterior parameters, we use the method proposed in \insertCite{ArtigokParametrico;textual}{kDGLM}.

For the details about the implementation see  \insertCite{ArtigoPacote;textual}{kDGLM}.
}
\examples{

# Normal case
T <- 200
mu <- rnorm(T, 0, 0.1)
data <- rnorm(T, cumsum(mu))

level <- polynomial_block(
  mu = 1,
  D = 0.95
)
variance <- polynomial_block(
  sigma2 = 1
)

# Known variance
outcome <- Normal(mu = "mu", Sigma = 1, outcome = data)

fitted_data <- fit_model(level, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, lag = -1)$plot

# Unknown variance
outcome <- Normal(mu = "mu", Sigma = "sigma2", outcome = data)

fitted_data <- fit_model(level, variance, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, lag = -1)$plot

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_model}}

Other {auxiliary functions for a creating outcomes}: 
\code{\link{Gamma}()},
\code{\link{Multinom}()},
\code{\link{Poisson}()},
\code{\link{report_distr}()}
}
\concept{{auxiliary functions for a creating outcomes}}
