% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.fitted_dlm}
\alias{plot.fitted_dlm}
\title{plot.fitted_dlm}
\usage{
\method{plot}{fitted_dlm}(x, ...)
}
\arguments{
\item{x}{A fitted_dlm object.}

\item{...}{Arguments passed to show_fit.}
}
\description{
plot method for class fitted_dlm
}
\keyword{internal}
