% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.fitted_dlm}
\alias{plot.fitted_dlm}
\title{plot.fitted_dlm}
\usage{
\method{plot}{fitted_dlm}(x, ...)
}
\arguments{
\item{x}{A fitted_dlm object.}

\item{...}{Arguments passed to show_fit.}
}
\description{
plot method for class fitted_dlm
}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lat_var}()},
\code{\link{plot_lin_pred}()},
\code{\link{print.fitted_dlm}()},
\code{\link{show_fit}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
\keyword{internal}
