% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{plot_lat_var}
\alias{plot_lat_var}
\title{Visualizing latent states in a fitted kDGLM model}
\usage{
plot_lat_var(
  model,
  var = "",
  lag = -1,
  cutoff = floor(model$t/10),
  pred_cred = 0.95,
  plotly = requireNamespace("plotly", quietly = TRUE)
)
}
\arguments{
\item{model}{fitted_dlm: A fitted DGLM model.}

\item{var}{Character: The name of the variables to plot (same value passed while creating the structure). Any variable whose name partially match this variable will be plotted.}

\item{lag}{Integer: A integer with the number of steps ahead should be used for evaluating the latent variables. If lag<0, the smoothed distribution is used and, if lag==0, the filtered distribution is used.}

\item{cutoff}{Integer: The number of initial steps that should be skipped in the plot. Usually, the model is still learning in the initial steps, so the estimated values are not reliable.}

\item{pred_cred}{Numeric: The credibility value for the credibility interval.}

\item{plotly}{Bool: A flag indicating if plotly should be used to create plots.}
}
\value{
A list containing:
\itemize{
   \item plot ggplot or plotly object: A plot showing the predictive mean and credibility interval with the observed data.
   \item data data.frame: A data frame containing the data used in the plot.
}
}
\description{
Visualizing latent states in a fitted kDGLM model
}
\examples{

T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)
summary(fitted_data)

plot_lat_var(fitted_data, smooth = TRUE)$plot

}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lin_pred}()},
\code{\link{report_dlm}()},
\code{\link{report_searched_dlm}()},
\code{\link{show_fit}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
