% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{plot_lin_pred}
\alias{plot_lin_pred}
\title{Visualizing linear predictors in a fitted kDGLM model}
\usage{
plot_lin_pred(
  model,
  pred = "",
  smooth = model$smooth,
  cut_off = 10,
  pred_cred = 0.95,
  plotly = requireNamespace("plotly", quietly = TRUE)
)
}
\arguments{
\item{model}{<undefined class> or list: A fitted DGLM model.}

\item{pred}{Character: The name of the linear predictors to plot (same value passed while creating the structure). Any predictors whose name partially match this variable will be plotted.}

\item{smooth}{Bool: A flag indicating if the smoothed distribution should be used. If false, the filtered distribution shall be used.}

\item{cut_off}{Integer: The number of initial steps that should be skipped in the plot. Usually, the model is still learning in the initial steps, so the estimated values are not reliable.}

\item{pred_cred}{Numeric: The credibility value for the credibility interval.}

\item{plotly}{Bool: A flag indicating if plotly should be used to create plots.}
}
\value{
A list containing:
\itemize{
   \item plot ggplot or plotly object: A plot showing the predictive mean and credibility interval with the observed data.
   \item data tibble: A data frame containing the data used in the plot.
}
}
\description{
Visualizing linear predictors in a fitted kDGLM model
}
\examples{

T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)
summary(fitted_data)

plot_lin_pred(fitted_data, smooth = TRUE)$plot

}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lat_var}()},
\code{\link{report_distr}()},
\code{\link{report_dlm}()},
\code{\link{report_searched_dlm}()},
\code{\link{show_fit}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
