% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.dlm_block}
\alias{+.dlm_block}
\title{+.fitted_dlm}
\usage{
\method{+}{dlm_block}(e1, e2)
}
\arguments{
\item{e1}{A dlm_block.}

\item{e2}{A dlm_block.}
}
\value{
The combined blocks as a dlm_block.
}
\description{
Define add operator for class dlm_block.
This method is wrapper for the block_superpos function.
}
\seealso{
\code{\link{block_superpos}}
}
\keyword{internal}
