% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dlm_block}
\alias{print.dlm_block}
\title{print.dlm_block}
\usage{
\method{print}{dlm_block}(x, ...)
}
\arguments{
\item{x}{A dlm_block object.}

\item{...}{Arguments passed to summary.dlm_block}
}
\value{
No return value, called to print a summary of a kDGLM structure.
}
\description{
This method is wrapper for the summary.dlm_block function.
}
\seealso{
\code{\link{summary.dlm_block}}
}
\keyword{internal}
