% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dlm_block}
\alias{print.dlm_block}
\title{print.dlm_block}
\usage{
\method{print}{dlm_block}(x, ...)
}
\arguments{
\item{x}{A dlm_block object.}

\item{...}{Arguments passed to summary.dlm_block}
}
\description{
print.dlm_block
}
\keyword{internal}
