% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.dlm_distr}
\alias{print.dlm_distr}
\title{print.dlm_distr}
\usage{
\method{print}{dlm_distr}(x, ...)
}
\arguments{
\item{x}{A dlm_distr object.}

\item{...}{Arguments passed to summary.dlm_distr}
}
\value{
No return value, called to print a summary of a kDGLM outcome.
}
\description{
This method is wrapper for the summary.dlm_distr function.
}
\seealso{
\code{\link{summary.dlm_distr}}
}
\keyword{internal}
