% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fitted_dlm}
\alias{print.fitted_dlm}
\title{print.fitted_dlm}
\usage{
\method{print}{fitted_dlm}(x, ...)
}
\arguments{
\item{x}{A fitted_dlm object.}

\item{...}{Arguments passed to summary.fitted_dlm}
}
\description{
print.fitted_dlm
}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot.dlm_coef}()},
\code{\link{plot.fitted_dlm}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
\keyword{internal}
