% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.searched_dlm}
\alias{print.searched_dlm}
\title{print.searched_dlm}
\usage{
\method{print}{searched_dlm}(x, ...)
}
\arguments{
\item{x}{A searched_dlm object.}

\item{...}{Arguments passed to summary.searched_dlm}
}
\description{
print.searched_dlm
}
\seealso{
Other {auxiliary visualization functions for the fitted.dlm class}: 
\code{\link{print.dlm_distr}()},
\code{\link{report_dlm}()},
\code{\link{report_searched_dlm}()},
\code{\link{summary.fitted_dlm}()},
\code{\link{summary.searched_dlm}()}
}
\concept{{auxiliary visualization functions for the fitted.dlm class}}
\keyword{internal}
