% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{report_block}
\alias{report_block}
\title{Summary for a kDGLM structure}
\usage{
report_block(dlm.block)
}
\arguments{
\item{dlm.block}{A dlm_block object.}
}
\description{
Prints a report for a dlm_block object.
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{*.dlm_block}()},
\code{\link{+.dlm_block}()},
\code{\link{AR_block}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{print.dlm_block}()},
\code{\link{regression_block}()},
\code{\link{summary.dlm_block}()}
}
\concept{{auxiliary functions for structural blocks}}
\keyword{internal}
