% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{report_distr}
\alias{report_distr}
\title{Summary for a kDGLM outcome}
\usage{
report_distr(dlm_distr)
}
\arguments{
\item{dlm_distr}{A fitted_dlm object.}
}
\description{
Prints a report for a dlm_distr object.
}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lat_var}()},
\code{\link{plot_lin_pred}()},
\code{\link{report_dlm}()},
\code{\link{report_searched_dlm}()},
\code{\link{show_fit}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
\keyword{internal}
