% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{report_distr}
\alias{report_distr}
\title{Summary for a kDGLM outcome}
\usage{
report_distr(dlm_distr)
}
\arguments{
\item{dlm_distr}{A dlm_distr object.}
}
\description{
Prints a report for a dlm_distr object.
}
\seealso{
Other {auxiliary functions for a creating outcomes}: 
\code{\link{Gamma}()},
\code{\link{Multinom}()},
\code{\link{Normal}()},
\code{\link{Poisson}()}
}
\concept{{auxiliary functions for a creating outcomes}}
\keyword{internal}
