% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{report_dlm}
\alias{report_dlm}
\title{Summary for a fitted kDGLM model}
\usage{
report_dlm(
  fitted_dlm,
  t = fitted_dlm$t,
  smooth = fitted_dlm$smooth,
  metric_cutoff = round(fitted_dlm$t/10)
)
}
\arguments{
\item{fitted_dlm}{A fitted_dlm object.}

\item{t}{Integer: The time index for the latent states.}

\item{smooth}{Bool: A flag indicating if the smoothed distribution should be used for the latent states.}

\item{metric_cutoff}{Integer: The cutoff time index for the metric calculation. Values before that time will be ignored.}
}
\description{
Prints a report for a fitted_dlm object.
}
\examples{

T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)
report_dlm(fitted_data)

}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lat_var}()},
\code{\link{plot_lin_pred}()},
\code{\link{report_distr}()},
\code{\link{report_searched_dlm}()},
\code{\link{show_fit}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
