% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{search_model}
\alias{search_model}
\title{Search the best hyper parameters fo a kDGLM model}
\usage{
search_model(
  ...,
  outcomes,
  search_grid,
  condition = "TRUE",
  smooth = TRUE,
  pred_cred = 0.95
)
}
\arguments{
\item{...}{dlm_block object: The structural blocks of the model. At least one block must be "undefined". See polynomial_block for more details.}

\item{outcomes}{List: The observed data. It should contain objects of the class dlm_distr.}

\item{search_grid}{List: A named list containing the possible values of each undefined hyper parameter.}

\item{condition}{Character: A character defining which combinations of undefined hyper parameter should be tested. See example for details.}

\item{smooth}{Bool: A flag indicating if the smoothed distribution for the latent variables should be calculated.}

\item{pred_cred}{Numeric: A number between 0 and 1 (not included) indicating the credibility interval for predictions. If not within the valid range of values, 0.95 will be used.}
}
\value{
A searched_dlm object containing the following values:
\itemize{
   \item search_data Data.frame: A table contain the metrics of each combination of hyper parameter tested and ordered from best combination to worst.
   \item structure dlm_block: The initial structure passed by the user, by with the undefined hyper parameters set to the best values found.
   \item model fitted_dlm: A model fitted with the best combination of hyper parameters.
}
}
\description{
Search the best hyper parameters fo a kDGLM model
}
\details{
This is an auxiliary function to search for the best combination of hyper parameters among the possible variations specfied by the user.
This function simple evaluates all possible combinations and chooses the one that results in the best fitting.
To compare each model, we evaluate the predictive log likelyhood for the one-step ahead prediction. We also compute the Relative Absolute Error (RAE) and return it to the user. It is important to note that one should be careful with the RAE metric, as it only takes into account the ponctual prediction.
}
\examples{

T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = "D_level")
season <- harmonic_block(rate = "sazo_effect", period = 40, D = "D_sazo")

outcome <- Poisson(lambda = "rate", outcome = data)

search_model(level, season,
  outcomes = outcome,
  search_grid = list(
    sazo_effect = c(0, 1),
    D_level = c(seq(0.8, 1, 0.05)),
    D_sazo = c(seq(0.95, 1, 0.01))
  ),
  condition = "sazo_effect==1 | D_sazo==1"
)

}
\seealso{
\code{\link{fit_model}}
}
