% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{set.block.value}
\alias{set.block.value}
\title{set.block.value}
\usage{
set.block.value(block, ...)
}
\arguments{
\item{block}{dlm_block: A undefined dlm_block object from which the undefined parameters shall be substituted.}

\item{...}{A set of named values for each unknown parameter.}
}
\value{
The initual block, but with the undefined parameters set to the chosen values.
}
\description{
Sets the values of undefined parameters in a block to those passed by the user.
}
\examples{

season <- harmonic_block(rate = 1, period = 12, D = "D.sazo") |>
  set.block.value(D.sazo = 0.975)
}
