% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{show_fit}
\alias{show_fit}
\title{Visualizing a fitted kDGLM model}
\usage{
show_fit(model, pred_cred = 0.95, lag = 1, plot_pkg = "auto")
}
\arguments{
\item{model}{fitted_dlm: A fitted DGLM.}

\item{pred_cred}{Numeric: The credibility value for the credibility interval.}

\item{lag}{Integer: A integer with the number of steps ahead should be used for prediction. If lag<0, the smoothed distribution is used and, if lag==0, the filtered distribuition is used.}

\item{plot_pkg}{String: A flag indicating if a plot should be produced. Should be one of 'auto', 'base', 'ggplot2' or 'plotly'.}
}
\value{
A list containing:
\itemize{
   \item data tibble object: A data frame containing the observations, predictions and credibility intervals at each time.
   \item plot ggplot or plotly object: A plot showing the predictive mean and credibility interval with the observed data.
}
}
\description{
Calculate the predictive mean and some quantile for the observed data and show a plot.
}
\examples{
T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, lag = -1)$plot

}
\seealso{
Other {auxiliary visualization functions for the fitted_dlm class}: 
\code{\link{plot_lat_var}()},
\code{\link{plot_lin_pred}()},
\code{\link{report_dlm}()},
\code{\link{report_searched_dlm}()}
}
\concept{{auxiliary visualization functions for the fitted_dlm class}}
