% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dlm_block}
\alias{summary.dlm_block}
\title{summary.dlm_block}
\usage{
\method{summary}{dlm_block}(object, ...)
}
\arguments{
\item{object}{A dlm_block object.}

\item{...}{Arguments passed to report_dlm}
}
\description{
summary method for class dlm_block
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{*.dlm_block}()},
\code{\link{+.dlm_block}()},
\code{\link{AR_block}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{print.dlm_block}()},
\code{\link{regression_block}()},
\code{\link{report_block}()}
}
\concept{{auxiliary functions for structural blocks}}
\keyword{internal}
