% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.dlm_block}
\alias{summary.dlm_block}
\title{summary.dlm_block}
\usage{
\method{summary}{dlm_block}(object, ...)
}
\arguments{
\item{object}{A dlm_block object.}

\item{...}{Arguments passed to report_dlm}
}
\description{
summary method for class dlm_block
}
\keyword{internal}
