% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary.dlm_distr}
\alias{summary.dlm_distr}
\title{Summary for a kDGLM outcome}
\usage{
\method{summary}{dlm_distr}(object, ...)
}
\arguments{
\item{object}{A dlm_distr object.}
}
\value{
No return value, called to print a summary of a kDGLM outcome.
}
\description{
Prints a report for a dlm_distr object.
}
\seealso{
Other auxiliary functions for a creating outcomes: 
\code{\link{Gamma}()},
\code{\link{Multinom}()},
\code{\link{Normal}()},
\code{\link{Poisson}()}
}
\concept{Reports for dlm_distr objects.}
\concept{auxiliary functions for a creating outcomes}
\keyword{internal}
