% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary.searched_dlm}
\alias{summary.searched_dlm}
\title{Summary for a searched_dlm object}
\usage{
\method{summary}{searched_dlm}(object, ...)
}
\arguments{
\item{object}{A searched_dlm object.}
}
\value{
No return value, called to print a summary of a searched_dlm object.
}
\description{
Prints a report for a searched_dlm object.
}
\seealso{
Other auxiliary visualization functions for the fitted_dlm class: 
\code{\link{plot.dlm_coef}()},
\code{\link{plot.fitted_dlm}()},
\code{\link{summary.fitted_dlm}()}
}
\concept{auxiliary visualization functions for the fitted_dlm class}
\keyword{internal}
