% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.searched_dlm}
\alias{summary.searched_dlm}
\title{summary.searched_dlm}
\usage{
\method{summary}{searched_dlm}(object, ...)
}
\arguments{
\item{object}{A searched_dlm object.}

\item{...}{Arguments passed to report_searched_dlm}
}
\description{
summary method for class searched_dlm
}
\seealso{
Other {auxiliary visualization functions for the fitted.dlm class}: 
\code{\link{print.dlm_distr}()},
\code{\link{print.searched_dlm}()},
\code{\link{report_dlm}()},
\code{\link{report_searched_dlm}()},
\code{\link{summary.fitted_dlm}()}
}
\concept{{auxiliary visualization functions for the fitted.dlm class}}
\keyword{internal}
