% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{*.dlm_block}
\alias{*.dlm_block}
\title{*.fitted_dlm}
\usage{
\method{*}{dlm_block}(e1, e2)
}
\arguments{
\item{e1}{A dlm_block (if e2 is an integer) or an integer (if e2 is a dlm_block).}

\item{e2}{An integer (if e1 is an dlm_block) or a dlm_block (if e1 is an integer).}
}
\value{
The combined replicated blocks as a dlm_block.
}
\description{
Define product operator for class dlm_block.
This method is wrapper for the block_mult function.
}
\seealso{
\code{\link{block_mult}}
}
\keyword{internal}
