% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{*.dlm_block}
\alias{*.dlm_block}
\title{*.fitted_dlm}
\usage{
\method{*}{dlm_block}(e1, e2)
}
\arguments{
\item{e1}{A dlm_block (if e2 is an Integer) or an Integer (if e2 is a dlm_block).}

\item{e2}{An Integer (if e1 is an dlm_block) or a dlm_block (if e1 is an Integer).}
}
\description{
Define product operator for class dlm_block
}
\seealso{
Other {auxiliary functions for structural blocks}: 
\code{\link{+.dlm_block}()},
\code{\link{AR_block}()},
\code{\link{block_mult}()},
\code{\link{block_rename}()},
\code{\link{block_superpos}()},
\code{\link{harmonic_block}()},
\code{\link{noise_block}()},
\code{\link{polynomial_block}()},
\code{\link{print.dlm_block}()},
\code{\link{regression_block}()},
\code{\link{report_block}()},
\code{\link{summary.dlm_block}()}
}
\concept{{auxiliary functions for structural blocks}}
\keyword{internal}
