% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multinormal.R
\name{update_multi_NG_correl}
\alias{update_multi_NG_correl}
\title{update_multi_NG_correl}
\usage{
update_multi_NG_correl(conj.param, ft, Qt, y, parms)
}
\value{
The parameters of the posterior distribution.
}
\description{
update_multi_NG_correl
}
\keyword{internal}
