% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{update_Normal}
\alias{update_Normal}
\title{update_Normal}
\usage{
update_Normal(conj.param, ft, Qt, y, parms)
}
\arguments{
\item{conj.param}{list: A vector containing the concentration parameters of the Normal.}

\item{ft}{numeric: A vector representing the means from the normal distribution. Not used in the default method.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution. Not used in the default method.}

\item{y}{numeric: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. For this kernel, parms should containing the covariance matrix parameter (V) for the observational Normal model.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Normal, assuming that the observed values came from a Normal model from which the covariance is known and the prior distribution for the mean vector have Normal distribution
}
\seealso{
Other auxiliary functions for a Normal outcome: 
\code{\link{convert_multi_NG_Normal}()},
\code{\link{multi_normal_gamma_pred}()},
\code{\link{normal_pred}()}
}
\concept{auxiliary functions for a Normal outcome}
\keyword{internal}
